added discount status

+++ src/main/java/mirkoabozzi/Abozzi/Market/dto/ReviewsDTO.java
@@ -9,7 +9,7 @@ public record ReviewsDTO(
         @NotNull(message = "Raring is required. ")
         @Min(value = 1, message = "Rating should not be less than 1 ")
         @Max(value = 5, message = "Rating should not be greater than 5 ")
-        Double rating,
+        int rating,
         @NotEmpty(message = "Comment is required. ")
         String comment,
         @NotEmpty(message = "Product is required. ")
